Switch on warnerrors and universal wheel
The warnerrors boolean option is used to tell Sphinx builders to
treat warnings as errors which will cause sphinx-build to fail if
it encounters warnings. This is generally useful to ensure your
documentation stays clean once you have a good docs build.

since we are python 2 and 3 compatible we should create a
universal wheel distribution.

Change-Id: I2a45596590d20b50abd232e3b070454cc28113ad