Use configmap to mount files
We should not use disk based mounts. Let's just use
configmap to load the files:
http://kubernetes.io/docs/user-guide/configmap/

NOTE: docker daemon should be started with
/var/lib/kubelet:/var/lib/kubelet:rw,shared
as mentioned in the following bug report for
the configmap based mounts to work.
https://github.com/kubernetes/kubernetes/issues/23392#issuecomment-209663941

We iterate over the file here and not just throw
the directory to be able to log exactly what we
picked up from the directories (and in the future
pull just specific files if necessary).

Implements blueprint openstack-services
Change-Id: I0fc8f19678234520c10bb14a661f6c8464b602ba