WIP - Implement a Replication Controller for Mariadb Bootstrap
Using a replication controller and etcd, we can signal to Kubernetes
when a bootstrapping job is in progress and when it's complete.

When a job is in progress, the pods associated with the service will
poll etcd waiting for mariadb_bootstrapped to return true and exit(1).
Kubernetes will reschedule the pods until it passes.

The bootstrap container will manage the variable by signaling etcd
that a bootstrap is in progress. After the boostrap is complete,
the job will scale be scaled down.

Change-Id: I06de8ebc347de3eab9aa58e8af85d9fa76182af5
Partially-implements: blueprint replication-controllers