Convert MariaDB to work without HostNetwork=True
MariaDB needs to be configured to listen to 0.0.0.0 in the Kolla config for this to work -- this is OK because we're only exposing the MariaDB port internally to the cluster.  That still needs to be done.

I ran some tests and HostNetwork=True doesn't allow for the k8s service discovery mechanism to work.  However, I also ran some tests and a container with HostNetwork=True is still able to connect to MariaDB's service -- Kubernetes creates the illusion of an SDN using iptables rules on the base node.  Thus, we're going to move faster on the k8s side if we default to HostNetwork=false and use HostNetwork=true when necessary (for example: Neutron).

Change-Id: I890247b908ff4592ed4f78c2eb369d26b3abfb6c
Partially-implements: blueprint kolla-kubernetes-service-discovery