The Keystone bootstrap job need to run a db sync
The Keystone bootstrap job was not running a database sync.
It was creating the users, passwords, and Keystone endpoints.

The bootstrap job was also failing to find /var/log/kolla/keystone.
We need to mount in /var/log/kolla/keystone so that the container
can change its permissions.

Change-Id: Ia4d781542a3f02d05c4287fd6f44e576303648e6
Partially-implements: blueprint openstack-services
Closes-bug: #1589285
Closes-bug: #1589284