Convert Memcached to work without HostNetwork=True
Memcached needs to be configured to listen to 0.0.0.0 in the Kolla
config for this to work -- this is OK because we're only exposing the
Memcached port internally to the cluster.  That still needs to be done.

I ran some tests and HostNetwork=True doesn't allow for the k8s
service discovery mechanism to work.  However, I also ran some tests
and a container with HostNetwork=True is still able to connect to
Memcached's service -- Kubernetes creates the illusion of an SDN using
iptables rules on the base node.  Thus, we're going to move faster on
the k8s side if we default to HostNetwork=false and use
HostNetwork=true when necessary (for example: Neutron).

Change-Id: I0858f78aac84c64e2640614b7e82f9c39de6daf7
Partially-implements: blueprint kolla-kubernetes-service-discovery
Partially-implements: blueprint remove-net-hostremove-net-host