Convert RabbitMQ to work without HostNetwork=True
RabbitMQ needs to be configured to listen to 0.0.0.0 in the Kolla
config for this to work -- this is OK because we're only exposing the
RabbitMQ port internally to the cluster.  That still needs to be done.

I ran some tests and HostNetwork=True doesn't allow for the k8s
service discovery mechanism to work.  However, I also ran some tests
and a container with HostNetwork=True is still able to connect to
RabbitMQ's service -- Kubernetes creates the illusion of an SDN using
iptables rules on the base node.  Thus, we're going to move faster on
the k8s side if we default to HostNetwork=false and use
HostNetwork=true when necessary (for example: Neutron).

I'm also exposing the rabbitmq management port inside of the cluster.

Change-Id: Ibd4f8d4d43f8b43023a45d9cd40fe74bbe535076
Partially-implements: blueprint kolla-kubernetes-service-discovery
Partially-implements: blueprint remove-net-hostremove-net-host