The Keystone bootstrap job need to run a db sync
The Keystone bootstrap job was not running a database sync.
It was creating the users, passwords, and Keystone endpoints.

The bootstrap job was also failing to find /var/log/kolla/keystone.
We need to mount in /var/log/kolla/keystone so that the container
can change its permissions.

Jenkins wasn't verifiing this patch. Moving the patch here.

Change-Id: Ia6550975a150ef412a4e8f4898f5f958bd842b00
Partially-implements: blueprint openstack-services
Closes-bug: #1589285
Closes-bug: #1589284