Horizon Versioning
Horizon is a web application that is distributed between the web
server, and the web browser. If the web server is upgraded while
the application is still running on the web browser, unusual issues
can creep up.

This patch allows multiple versions of Horizon to be run on
Kubernetes in parallel. Further patches will use this to direct
existing browser sessions at the version of Horizon they started on
to allow seamless major rolling upgrades.

Change-Id: Ib1d840aa10f02dfec6c89b7fa8a4fcf0ea204994