Horizon Frontend
Horizon is a web application that is distributed between the web
server, and the web browser. If the web server is upgraded while
the application is still running on the web browser, unusual issues
can creep up.

This patch allows for an haproxy based frontend container to be
created that, through kolla config options, allows multiple horizon
backends to be defined, browsers to be temporarily bound to the
versions they initially landed on, and new horizons introduced for
testing and promotion to default.

Intended to be used with: Ib1d840aa10f02dfec6c89b7fa8a4fcf0ea204994

Change-Id: I8ce7d89f42594a2b0617f1a9da666da351e3f694
Depends-On: I2dcd0fc224a959cc36cb7fd2a2adc99be280b7ff