POC: Kolla-Kubernetes multi-node persistence for Mariadb
Note: This is still a proof of concept until I can create a patch to
fix the legacy "kolla-kubernetes bootstrap" command which does a
"kubectl create -f <directory>" on all bootstrap files for a
particular service at the same time.  This worked fine before, because
the templates in the bootstrap directory could be instantiated without
regard to order.  However, since we have added disk, pv, and pvc to
the bootstrap directory, we need a very specific order.  Once I create
that patch, I'll rebase this patch and it should be ready for merge.
The real commit messages follows:

This patch uses the Mariadb service to show how we may support the
configuration and use of kubernetes persistent volumes across storage
providers, including Ceph, AWS, GCE, and host development mounts on
the all-in-one Hyperkube.

Problem Statement: Currently, boostrap and running of kolla containers
requiring persistence will fail on multi-node kubernetes clusters
because the bootstrap using a hostmount may occur on a different node
than the node on which the service pod will run.  Thus, in order to
support multi-node deployments of kolla-kubernetes, we must support
persistent volumes that replication controllers can associate with and
follow pods which may move from one host to another.

See ./doc/source/multi-node.rst for the multi-node deployment guide.

Change-Id: I32968cd4faed5413066af28c25ca32b6e79743ac