WIP Adds ansible support to deploy skydns service
Discovering IP addresses of OpenStack serviceis especially in a
multinode environment is required, hence we have to deploy skydns service.
Relaying on an operator to deploy skydns or another service proving
similar functionality is not relable option as kolla-kubernetes is not aware
how it was configured. Deploying skydns as a part of deploying
kolla-kubernetes is much more realiable option and should be prefered.

Change-Id: I409d120020069246b1e7d0dbc22adaeb349078fc
Partially-implements: blueprint ansible-orchestration
Partially-Implements: blueprint kolla-kubernetes-service-discovery