WIP(do not merge) Add status command for Workflow Engine
Added a status command:
  kolla-kubernetes resource status svc keystone

The above will print out a yaml status dict of kube resources defined
by the resource_type.  The workflow engine may query the status of any
service from the perspective of kubernetes (e.g. is the resource
created, are the replication controllers in good states, etc).  The
output of this state is yaml, and the workflow engine may key off of
the top level "status: (ok|error)".

The file ./etc/kolla-kubernetes/services-resources.yml defines a list
of services and for each service a list of resource_types and a list
of files under each resource type.  In the keystone example below,
executing "kolla-kubernetes resource create svc keystone" will process
both of the files under the resource_type svc and send the output to
"kubectl create -f -" to create the resources.

  - name: keystone
    pods:
    - name: keystone
      containers:
      - name: keystone
    resources:
      disk:
      pv:
      pvc:
      svc:
      - services/keystone/keystone-service-admin.yml.j2
      - services/keystone/keystone-service-public.yml.j2
      bootstrap:
      - services/keystone/keystone-bootstrap-job.yml.j2
      pod:
      - services/keystone/keystone-pod.yml.j2

Any of the template files, once jinja processed, may generate a stream
of kube yaml resource definitions separated by "\n---\n".  This code
will first generate the template for a particular file, and then
individually feed each event to "kubectl describe -f -", while parsing
the output and executing checks based on the type of input.  For
example, for ReplicationControllers it checkes if desired_replicas ==
current_replicas.  Then, it sets the state (ok|error) of each segment,
which rolls up to the state of each template file, which rolls up to
the state of the resource_type for the service.  The resulting yaml is
printed to stdout as yaml, at which point operators may view the
result or workflow systems may key off of values such as the top-level
key "status: (ok|error)"

Sample Output:

$ kolla-kubernetes resource status svc keystone
"files":
- "errors": &id003 []
  "file": "/home/vagrant/Dev/kolla-kubernetes/services/keystone/keystone-service-admin.yml.j2"
  "ok": &id004 []
  "segments":
  - "errors": &id001 []
    "ok": &id002
    - "Verified resource with name keystone-admin exists"
    - "Service has IP"
    - "Verified resource with name keystone-public exists"
    - "Service has IP"
    "status": "ok"
  - "errors": *id001
    "ok": *id002
    "status": "ok"
  "status": "ok"
- "errors": *id003
  "file": "/home/vagrant/Dev/kolla-kubernetes/services/keystone/keystone-service-public.yml.j2"
  "ok": *id004
  "segments":
  - "errors": *id001
    "ok": *id002
    "status": "ok"
  - "errors": *id001
    "ok": *id002
    "status": "ok"
  "status": "ok"
"resource_type": "svc"
"service_name": "keystone"
"status": "ok"

Change-Id: Id3077b71800d348cc674d000104727a652ebfb83