Deployment, Readiness, Safe Shutdown, & Scaling for glance-api
This patch switches glance-api over to use the k8s deployment type.
It also puts an haproxy in front of it so it can safely be drained
to perform rolling upgrades/node migrations and scale down. It
blocks access until the service is up via readiness probe. Lastly,
it removes an unneeded volume when ceph direct access is configured.

Change-Id: Id95c2f0d891882a6cc91a8248156c998a204a294