WIP- openvswitch ip
The neutron openvswitch agent needs to know the ip address of the
tunnel interface to operate properly.

Also splits/renames the daemonset to be more easily found and sharable
with the network node, and allows the openvswitch containers to be
excluded so that if you are running openvswitch on the host you can
reuse it.

Change-Id: I0691a2286a4b89720066c13b628b7221eb9422c5