Checks to start reducing Technical Debt
We already have a lot of technical debt. This test adds checks for 3
of the types we currently have. There will be follow up patches for
adding tests for more of the types later. The tests this implements
are:
1. Not having explicit namespaces on namespaced resources.
2. Having namespaces in the template description. Its implied by the
   namespace in the deployment/petset/daemonset/replicationcontroller
3. Not having a nodeSelector on a resource.

It also white lists all existing occurrences so they can be picked
off in followup patches and tested for correctness.

Change-Id: I316f8c3edf24d33156e8effc23a825c15a158cdb