Cleanup container names
This ensures that all containers that have a main container are
named 'main' to make it easier on operators to enter them or get
logs from them. The Pod name stays unique so there will never
be conflicts.

This also modifies the 'main' container check to ignore Jobs
with multiple containers. A few Jobs don't really have a main.

Change-Id: Ia10718141064752d1aea280dc7663a3b0ed87dfa
Implements: blueprint container-names