Break neutron endpoints job up further
This lets you run each job one at a time. It uses templates to do
it so it isn't much more code. This is a step towards being able
to move most of the bootstrap jobs into services/common too as there
is a lot of common code between the services.

Change-Id: Icec23ea533835808bcf5b4b39e26c0ab98f7216a