First stab at minikube doc
This document is the start of a procedure for bringing up
kolla-kubernetes from scratch on a dev box that only has
linux/windows/macosx installed. It pulls prebuilt containers
from the docker hub.

Note, it currently pulls a couple of containers from the
kfox1111 docker hub account until the containers can be
merged back into kolla and kolla has a 3.0.0 release.

TrivialFix

Change-Id: I2513fa374f1a58725ed321da97d96f2191acb0c5