CLI rework.
This Patch Set reworks the cli. For now, instead of replacing the
'kolla-kubernetes' tool it makes a new tool called 'kollakube'.
This new command requires less typing. Operators who will need to
type the command a lot, so shortening the cli will help over time.
In a future PS the old 'kolla-kubernetes' command will be removed.

It reworks the resource and template system to allow the operator
to specify multiple entries to be processed at once. This aligns
kollakube with kubectl behavior.

The 'resource-template' subcommand is renamed to just 'template'.

The 'resource-map' subcommand is renamed to just 'map'.

The 'template' subcommand drops the action argument as it has no
effect.

Aliases are added from 'resource' to 'res' and 'template' to 'tmpl'.

The 'kollakube' cli removes any workflow related subcommands as they
are unlikely to work well moving forward and will be replaced
shortly with an ansible workflow.

Change-Id: Ide7b5d391a4ac0a514c027cf761033d9eb94d152