More services to deployments
This patch set changes nova-api and neutron-server to use Deployments
similar to the rest. This allows them to seamlessly scale and minor
upgrades to be applied without downtime.

It also makes some minor tweaks to the other Deployments to make them
 more similar to each other in preparation for a future PS to merge
them into a single template. They now look very similar.

Lastly, nova-novncproxy was in dos format. This changes it to be in
unix format like the rest.

Change-Id: I8ac7e938842b9aa48d536b4024368fdf82f7ff62
Partially-Implements: blueprint api-termination
Partially-Implements: blueprint deployments