Basic ceph deployment support
This PS adds support for scalable glance backed by ceph. It also
adds an implementation of ceph in k8s that can be used for testing
when no other ceph option is available. The k8s managed ceph is
intended mostly just for testing with minikube. Some groundwork is
also laid for nova and cinder support.

Change-Id: I5332c5b5ee4d6d1ec90cec6abffe88f20fb1373d