Split endpoint jobs and start testing the deployments
The previous check tested if services deploy. This adds testing to
see what it deployed works. It picked up an issue with the keystone
endpoint creation jobs which is fixed in this PS by splitting the
jobs into one job per task and sharing the templates between all
services. There seems to be an issue with ceph still sometimes,
which will be addressed in another PS.

Partially-Implements: blueprint job-templates

Change-Id: I83f480a043406ae1b1497060a865cbba45f05578