Kolla-Kubernetes Vagrant development environment documentation
This is very much a work in progress, but describes the basic
setup of a vagrant/ansible based development environment for
kolla kubernetes. Based on halcyon-kubernetes, it uses kubeadm
to set up a kubernetes cluster running either CentOS or Ubuntu
with helm, ceph mounting support on the hosts, and host access
to the k8s dns service.

Change-Id: I472fc8f02f3d56d170d1988f611a220d9232a8ef