WIP: Add Development Setup Script for use with Kubernetes Helm
This ps is just so I can move my work home - please review
the previous ps.

This brings in a helper script for setting up a development
environment for building and deploying Kolla-Kubernetes locally.

This rev fixes a few issues in the devenv setup that were blocking
me (sdake) during bringup.  I am currently stuck on an iscsi error.  I'll
paste the output today of the iscsi error in the review - perhaps someone
has some ides on how to resolve it.

Change-Id: I87e7a722ae611154311f330106023b3adfb629d0
Co-Authored-By: Kevin Fox <Kevin.Fox@pnnl.gov>