More dns gate fixes
The previous ps attempted to check for dns issues. We discovered some
more and this fixes the rest. The gate now checks to see if dns both
internally and externally works and blocks until it does. It fixes
external dns on Ubuntu. It also cleans up the gate output to make it
easier to see the last thing that failed.

Change-Id: I52226ff4d8419b717487428fbf0e5c832ceedb27