Fix python tox -e cover failure
Currently, running ``tox -e cover`` errors out and does not generate
the correct coverage report.  The problem is caused by (a) a typo in
``tox.ini``, and (b) a mix up between ``kolla_kubernetes`` vs.
``kolla-kubernetes`` in .coveragerc and setup.cfg.  This causes the
coverage report to not properly pick up the module. This patch set
addresses the problem.

Change-Id: I0a10c0eb563968f4b88a816276455a6afcf5df80
Closes-Bug: #1650266