Support rolling upgrades of horizon
Horizon is a web application that is distributed between the web
server, and the web browser. If the web server is upgraded while
the application is still running on the web browser, unusual issues
can creep up.

This patch set allows multiple versions of Horizon to be run on
Kubernetes in parallel. It adds a Horizon frontend with session
tracking that allows the browser to continue to use the Horizon
version it was started with.

Change-Id: Ib1d840aa10f02dfec6c89b7fa8a4fcf0ea204994