Pull container tallballs into the gate
Pull containers from tarballs.oo to make gate reliable.  This will
not fix rax.* problems - not sure what the deal is with the rax
providers.

This patch works by downloading the regitry from tarballs.oo, then
using docker lod to load in the tarballs.  Note we use wget in this
patch instead of curl as wget is far more reliable.

Change-Id: Ibc92755f20c4f4fcc1a0e13d7b2d58967a9dd5e9