DNM: Temporary Hack to enable pathfinder
The kolla-kube res create tool file pathfinder is
defective as a result of the repository repo split.  As
we intend to entirely remove kollakube from kolla-kubernetes
RSN, it doesn't make alot of sense to rework the gate to fix
this one problem.  Unfortunately this change is required to
actually get a valid deployment.  So that this patch doesn't
have to constantly be rebased, it is being pulled out from
any gate work.

Change-Id: I04d57d3c228da15a0aecab8f34d92f9a6a2486ff