Fix a bug related to genconfig and iscsi
ansible-playbook was not called in one place - instead genconfig
was being called.  This exposed a problem where iscsi was not in
the site.yml file.  It further exposed a problem where the change
Id1b1cc3bb3c1790a77f9449ec5b0b5d1c67f91e4 was not applied.

Co-Authored-By: zhubingbing<zhu.bingbing@99cloud.net>
Change-Id: I0e731b233ab6b4d8182ae811e16c1aa398f40f39
Partially-Impelments: move-config-to-kolla-k8s