A start at a simplified devtool environment
We have wasted *TWO* months or more of time attempting to integrate
the development environment with the main gate.  Instead of doing
that, lets start small with a set of scripts based upon the
deployment guide which was built based upon learning how the
gate operates.

The intent is to follow this up with a gate specifically for the
development environment workflow.  I know this is far from perfect.
THe idea is for folks to build on this work in an incremental fashion.
As such, I'd ask for a little liency in the review process.

This will ultimatelly integrate with
1ac7a13cc8cd092d5d75542d3dc7992d2f972dae

Thanks
-steve

Change-Id: I90671b4894f0024d3b094c8459edf61e1acfe04a
Partially-Implements: blueprint devenv-taketwo