Add a destroy workflow for kolla-kubernetes
This workflow deletes all of the configmaps, secrets, and helm charts
within the kolla namespace.  This workflow further unlabels all nodes.

After that is done with, the cluster should be back to its original state.

The intent of this patch is to make this workflow available for consumption
by a container.

Implements: blueprint workflow-destroy
Change-Id: I92f50b421281b2bd8bdbcf9cd771f849ebfe7b3c