Remove pathfinder from call path of secret-generator
Pathfinder is full of fail especially when attempting to remove
the pip install kolla-ansible operation.  To help that effort along
make secret-generator.py not depend on kubernetes_resources.yml which
really is overkill to specify a namespace.  For the moment hardcode
the namespace as this tool will likely be refactored into something
more helm-like.

Change-Id: Icec93b4c9162cb54956f5c48ea576bb360ad4458