Script to Deploy Kolla Kubernetes AIO
This is an attempt to consolidate the codes in the Gates so
that a developer can get a running Kolla Kubernetes cluster
on a CentOS 7 VM by executing the script.

There is no special requirements for the CentOS 7 VM. The
developer will install the CentOS using the minimal CentOS 7
disk.

The flow of the installation is similar to the bare metal
deployment guide.  The developer will need to provide 5
parameters to the script as that will be fed to the cloud.yaml

1) kube proxy IP
2) tunnel_interface
3) ext_interface_name
4) keepalived VIP
5) subnet size of management subnet

We can execute the script by running the below command (note
that values vary from environment to environment):

tools/setup_kube_AIO.sh 192.168.122.241 eth0 eth1 192.168.122.254 24

Change-Id: I5ccb32d2174bb2403fd69f28f07aae1b3d1ec454