Run ironic-dnsmasq in privileged mode
In order for dnsmasq unicast DHCPOffers to work correctly inside a
container it needs to modify the host's arp cache. The arp cache is part
of the host's network stack, running in net=host mode allows the
container to use the stack but not modify it. In order to modify it we
must also run this container in privileged mode.

Change-Id: I4a5dfa44858bd223960dc3af50a26843c31b2763