Combine ironic-inspector deployments into one pod
For the ironic-inspector service to operate correctly it must run on the
same host as its TFTP and DNSmasq service. This is because the
ironic-inspector service needs to manage iptables for DNSmasq to prevent
node that are not in inspection from receiving DHCP responses from
DNSmasq. To achieve this, we should combine the containers providing
those services into the same deployment as the ironic-inspector service
to ensure they are always deployed together.

Change-Id: Ib3af723a2c8b2cffb6eacec6293c5ea7860ba72f