Remove all kolla-ansible dependencies
See ticket https://bugs.launchpad.net/kolla-kubernetes/+bug/1693281.
The kolla-ansible project has deprecated ml2_conf.ini for several of the
Neutron microservices. This is a breaking change with the
kolla-kubernetes deployment guide. This commit updates the guide to
remove all kolla-ansible dependencies and installs the single
kolla-ansible tool required for the baremetal deployment to work.