Add option to enable ironic inspector LLDP collection
In order for LLDP information to be collected by ironic inspector we
have to pass a flag to enable it in the pxe kernel parameters. This
patch adds a new option that defaults to false, but when set to true in
clouds.yaml will append the flag to the end of the kernel parameters in
pxelinux.default.

Change-Id: I64e04714f83947cdefaacf38b0a4839c1d35ab5c