Add ironic provision network IP addresses to API external IPs
For the ironic baremetal servers to call back to the Ironic API they
need to know where to talk to, to make this HA we should ensure that
each conductor exposes the ironic API on its provisoning network IP
address. To do that this patch allows to specify a list of the conductor
provision network IPs in clouds.yaml so we can add them to the
ironic-api-svc external IPs list.

Change-Id: I4566f6eccf4d74240f0e96e420b285baa2bf06ee