[WIP] Add support for HTTPS on external API network
This patch enables the use of HTTPS for the OpenStack API by
using HAProxy to provide TLS termination.  Both the configuration
and the documentation for the deployment guide are updated with
this optional feature.

This patch is still in progress.

Next step is to update the Keystone catalog entries with https://
instead of http:// for those services presenting with TLS.

For the initial patch, only the services that currently use
HAProxy will support TLS.

Change-Id: I2ee6a379c424b2aaa7780dc412211496afe8f5b1