Moved ko.py to orchestration/tools
Python tool to bring up Kubernetes and OpenStack

This tool provides a method to deploy OpenStack on a Kubernetes
Cluster using Kolla and Kolla-Kubernetes on bare metal servers
or virtual machines. Virtual machines supported are Ubuntu and
Centos.

Versions supported are 4.x (Ocata) and 5.x.

I started developing it as I worked on the project and the
Deployment Guide.

In no way is this supposed to be an official production method.

It is however very simple to use (ko.py eth0 eth1), and I have
verified it on Centos and Ubuntu and Ocata and Pike.

It comes with a 'demo' mode that attempts to explain each step
to the user.

It comes with lots of switches to play with different versions
of all the tools used.

Change-Id: I85858d06efcbda686bcfa177d41e6f002f98cb03