Add the checking for nova-libvirt start
If use ceph as the backend, user need to copy the secrets
to the /etc/libvirt. If not use the ceph backend, this process
will not occur. So add the checking here.
The fix just solve 1 issues when booting nova-libvirt. Will need
more to fix the libvirtd could not start issue as following.

Partial-Bug: #1733806

Change-Id: Ief6501425bbea4b9128154b5a1193b0a2c4a6dd8
Signed-off-by: Kevin Zhao <kevin.zhao@arm.com>