Add architecture check when downloading helm and kubectl
There is an issue that setup-kubectl.sh and setup-helm.sh can only
download helm and kubectl for X86_64, but it cannot handle other
architectrues like arm64 or ppc64le.
This patch detects architecture and make the script architecture-aware,
so the script can download proper helm and kubectl for the corresponding
arm64 and ppc64le architectures, in addition to X86_64.

Closes-Bug: #1738743

Change-Id: I58a39268040797fdb3e39b9ff10b2cd7e0822386
Signed-off-by: Kevin Zhao <kevin.zhao@arm.com>