Add the flannel driver in test case
Add the flannel network driver in test case when
the Kolla-kubernetes test is not running on amd64
platform. Also delete part of the code that works around
the issue of lacking net-conf.json, since the current version
of canal.yml and kube-flannel.yml all have net-conf.json
now.

Closes-Bug: #1738317

Jira: ENTWLS-516

Change-Id: I0f7a90c26bdce7096f1501d48927d7c8a49886b5
Signed-off-by: Kevin Zhao <kevin.zhao@arm.com>