Setting up Cells V2
"Table 'nova_api.cell_mappings' doesn't exist" error occurs when starting nova-api-manage-db pod,
and nova-api-create-cell pod is stuck in "Init:2/3" status as the result of "openstack hypervisor list" is empty.
I think this is because cell databases are not properly created/registered/populated.
I added the below commands to nova-api-manage-db-job
  - nova-manage api_db sync;
  - nova-manage cell_v2 map_cell0;
  - nova-manage cell_v2 create_cell --name=cell1 --verbose;
  - nova-manage db sync;

Change-Id: I1c7c8c541d954d806382ed21fb8ad68cf3242ada
Closes-Bug: #1766753